General fission Info (Revision 2)
Added a lot of information to "General Info" in "Nuclear Fusion". I originally intended for the section on heatsinks to be in the now deleted "Solid Fuel" section, but I decided against this, since the behavior of the MSR coolant heaters are derivative from their solid heatsink counterparts. I figure that since solid fuel reactors are the baseline, it isn't worth having a special page for them, considering much of their behavior is universal. When MSRs and Heat Exchangers are fully implemented and mostly stable, it will be worth writing documentation for them, but for now I'm going to let them go. Subject to change, disscusion, and reversion.