>:(
I am extremley unsatisfied with this, but I decided it would be for the best to not to go in depth, and instead refer to the video. Perhaps making a speadsheet fully explaining all the math behind this, and then linking to that instead would be the best option. Subject to change, disscusion, and reversion.